+++ packages/giu-examples/pages/index.js
@@ -36,7 +36,6 @@ import {
   Hints,
   hintDefine,
   hintShow,
-  hoverable,
   flexContainer,
   isRequired,
   isEmail,
@@ -241,9 +240,6 @@ const Contents = ({ id }) => {
     case 'demo:progress':
       el = <ProgressDemo />;
       break;
-    case 'demo:hoverable':
-      el = <HoverableDemo />;
-      break;
     default:
       el = (
         <div>
@@ -294,7 +290,8 @@ class InputIntroAge extends React.Component {
                 title: 'Saved!',
                 msg: `Your age is ${age}`,
               })
-            )}
+            )
+          }
         >
           <Icon icon="save" /> Save
         </Button>
@@ -337,11 +334,7 @@ class InputTypes extends React.Component {
           </li>
           <li>
             <a href="#select">Select</a>:{' '}
-            <Select
-              items={getExampleItems()}
-              value="blueberries"
-              required
-            />{' '}
+            <Select items={getExampleItems()} value="blueberries" required />{' '}
             (native){' '}
             <Select
               required
@@ -359,7 +352,9 @@ class InputTypes extends React.Component {
           </li>
           <li>
             <a href="#colorinput">ColorInput</a>:{' '}
-            <ColorInput value="aadc54ab" /> (<a href="#color-inputs">more details</a>)
+            <ColorInput value="aadc54ab" /> (<a href="#color-inputs">
+              more details
+            </a>)
           </li>
           <li>
             <a href="#fileinput">FileInput</a>: <FileInput />
@@ -988,7 +983,8 @@ class DropDownMenus extends React.Component {
               msg: `I'll fetch some ${fruits}!`,
               icon: 'shopping-cart',
               type: 'success',
-            })}
+            })
+          }
           style={{ padding: '3px 8px' }}
         >
           <Icon icon="shopping-basket" />&nbsp; Which fruit?
@@ -1117,7 +1113,8 @@ const Buttons = ({ fIncludeDisabled }) => (
         createNotif({
           icon: 'cutlery',
           msg: 'What about some sushi?',
-        })}
+        })
+      }
     >
       <Icon icon="lightbulb-o" /> Inspire me!
     </Button>&nbsp;&nbsp;
@@ -1127,7 +1124,8 @@ const Buttons = ({ fIncludeDisabled }) => (
           type: 'success',
           icon: 'heart',
           msg: "You're welcome!",
-        })}
+        })
+      }
       plain
     >
       OK, thanks!
@@ -1207,20 +1205,6 @@ class ProgressDemo extends React.Component {
     );
   }
 }
-
-const HoverableDemo = hoverable(({ hovering, onHoverStart, onHoverStop }) => (
-  <LargeMessage
-    onMouseEnter={onHoverStart}
-    onMouseLeave={onHoverStop}
-    style={{
-      border: '1px solid #ddd',
-      cursor: 'default',
-      backgroundColor: hovering ? '#ddd' : undefined,
-    }}
-  >
-    Hover me {hovering}
-  </LargeMessage>
-));
 
 // -----------------------------------------------
 // Helpers